# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger документация проекта <<В добрые руки>>",
        "description": "Open API",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://in-good-hands.dev.mind4.me",
            "description": " API Server"
        }
    ],
    "paths": {
...
        "/api/send_bid": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Отправить заявку на бронирование",
                "operationId": "f7255b4d012ca6de9b0af918cfd23bd0",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_post",
                                    "days"
                                ],
                                "properties": {
                                    "id_post": {
                                        "type": "integer",
                                        "example": 111
                                    },
                                    "days": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/get_bids": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Получить заявки на бронирование (входящие или исходящие)",
                "operationId": "342c3cff35a80cef2894fceda5b3f427",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "example": 3
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "example": 1
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "required": true,
                        "example": "outcoming"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/confirm_bid": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Подтвердить заявку на бронирование",
                "operationId": "1418b3f7a634a14c7b91881c22c52c49",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_bid"
                                ],
                                "properties": {
                                    "id_bid": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/delete_bid": {
            "delete": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Отклонить\\удалить заявку на бронирование",
                "operationId": "d097a8d5ffbda8d6b1db42ca1a1e31f2",
                "parameters": [
                    {
                        "name": "id_bid",
                        "in": "query",
                        "required": true,
                        "example": 9
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://in-good-hands.dev.mind4.me/oauth/authorize",
                        "tokenUrl": "https://in-good-hands.dev.mind4.me/oauth/token",
                        "refreshUrl": "https://in-good-hands.dev.mind4.me/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}